requirements.txt
python-telegram-bot==20.7
aiodns==3.1.1
dnspython==2.4.2 